madurez {inbio}	R Documentation

Madurez Sexual en Talla y Edad

Description

Calcula la talla y la edad de madurez sexual aplicando un glm con errores binomiales. Mediante bootstrap no parametrico estima los coeficientes de variacion de: los parametros de la funcion, la L50 y la A50.

Usage


madurez(especie = "nombre especie", cl = 1, unid = "cm", edad = T, sex = F, n = 1000)

Arguments

especie
Nombre del archivo donde estan los datos de entrada, sin la extension y sin el directorio.

cl
Intervalo de talla de la especie. El valor por defecto es 1.

unid
Unidad de los datos de talla: cm o mm. Por defecto unid="cm"

edad
Si edad=T, se calcula la edad de madurez. Si edad=F, el proceso no calcula la edad de madurez. Por defecto edad=T.

sex
Si sex=F, el proceso analiza los datos de machos y hembras conjuntamente. Si sex=T, se analizan separadamente machos y hembras. Por defecto sex=F.

n
Numero de replicas bootstrap. El numero recomendado es 1000.

Value

Dos archivo "txt" con los resultados de los datos originales y de las replicas bootstrap y su coeficiente de variacion,uno para madurez en talla y otro para madurez en edad.

Dos archivos "pdf" con las graficas del ajuste con datos originales y bootstrap, las graficas del analisis de residuos del ajuste de los datos originales y analisis de las replicas bootstrap, un archivo para madurez en talla y otro para madurez en edad.

Si el proceso se realiza separadamente para machos y hembras, los archivos de salida tambien son independientes.

Author(s)

Paz Sampedro

References

Davison, A.C. & D.V. Hinkley. 1999. Bootstrap methods and their application. Cambridge University Press, 575 pp.

Efron, B. 1979. Bootstrap methods: another look at the jackknife. Annals of Statistics, 7. 1-26 p.

Roa, R., Ernst, B & F. Tapia. 1999. Estimation of size at sexual maturity: an evaluation of analytical and resampling procedures. Fishery Bulletin 97: 570-580.

Examples


## Requiere que exista el archivo "lepwhi2004.csv" en el directorio de trabajo
## Calcula la talla y la edad de madurez sexual y su coeficiente de variacion conjuntamente para ambos sexos
#   madurez (especie="lepwhi2004", sex=F)

## Requiere que exista el archivo "cangrejoreal.csv" en el directorio de trabajo
## Calcula solo la talla de madurez sexual (edad=F) y su coeficiente de variacion separadamente para machos y
## hembras. La talla es en milimetros.
#   madurez(especie="cangrejoreal", unid="mm", edad=F, sex=T)

[Package inbio version 3.0.0 Index]