peso {inbio}	R Documentation

Relacion Talla-Peso

Description

Calcula la relacion talla-peso por estimaci?n no linear y ajuste de m?nimos cuadrados. Mediante bootstrap no parametrico estima el coeficiente de variacion de los parametros de la relaci?n talla-peso.

Usage


peso(especie = "nombre especie", cl = 1, unid = "cm", sex = F, a = 5e-04, b = 3, n = 1000)

Arguments

especie
Nombre del archivo donde estan los datos de entrada, sin la extension y sin el directorio.

cl
Intervalo de talla. El valor por defecto es 1.

unid
Unidad de los datos de talla: cm o mm. Por defecto unid="cm".

sex
Si sex=F el proceso analiza los datos de machos y hembras conjuntamente. Si sex=T el proceso analiza separadamente machos y hembras. Por defecto sex=F.

a
Valor de partida ("start value" o "semilla") de la ordenada.

b
Valor de partida ("start value" o "semilla") de la pendiente.

n
Numero de replicas bootstrap. El numero recomendado es 1000.

Details

Es importante que los valores de partida ("start value" o "semilla") no sean muy diferentes de los que se quiere calcular (se pueden consultar valores de otros anhos o zonas), si no, el proceso iterativo por el que se realiza el ajuste podria fallar.

Value

Archivo "txt" con los resultados de los datos originales y de las replicas bootstrap, asi como su coeficiente de variacion.

Archivo "pdf" con las graficas del ajuste para los datos originales y para las replicas bootstrap, graficas del analisis de residuos del ajuste de los datos originales y graficas de analisis de las replicas bootstrap.

Si el proceso se realiza separadamente para machos y hembras, los archivos de salida tambien son independientes.

Author(s)

Paz Sampedro

References

Efron, B. 1979. Bootstrap methods: another look at the jackknife. Annals of Statistics, 7. 1-26 p.

Examples


## Requiere que exista el archivo "merluza.csv" en el directorio de trabajo
## Calcula la relacion talla-peso separadamente para machos y hembras
#   peso (especie="merluza", sex=T, a= 0.005, b=3)

## Requiere que exista el archivo "choco.csv" en el directorio de trabajo
## Calcula la relacion talla-peso para una especie con intervalo de talla de 0.5 cm, separadamente por sexos
#   peso(especie="choco", cl=0.5, sex=T, a=0.05, b=2.4)
 

[Package inbio version 3.0.0 Index]