sexratio {inbio}	R Documentation

Sexratio por Talla y Edad

Description

Calcula el porcentaje de hembras por talla y por edad, y su coeficiente de variacion mediante bootstrap no parametrico. ** Sexratio= (hembras/(hembras+machos)

Usage


sexratio(especie = "nombre especie", cl = 1, unid = "cm", edad = T, n = 1000)

Arguments

especie
Nombre del archivo donde estan los datos de entrada, sin la extension y sin el directorio.

cl
Intervalo de talla usado en la sexratio por talla. El valor por defecto es 1.

unid
Unidad de los datos de talla: cm o mm. Por defecto unid="cm".

edad
Si edad=T la rutina calcula el sexratio por edad; Si edad=F la rutina no calcula sexratio por edad. Por defecto edad = T.

n
Numero de replicas bootstrap. El numero recomendado es 1000.

Value

Archivo "txt" con los resultados de sexratio por talla y su coeficiente de variacion global y los resultados para sexratio por edad y coeficiente de variacion global.

Archivo "pdf" con dos graficas, una grafica del sexratio y coeficiente de variacion por talla y otra del sexratio y coeficiente de variacion por edad.

Author(s)

Paz Sampedro

References

Efron, B. 1979. Bootstrap methods: another look at the jackknife. Annals of Statistics, 7. 1-26 p.

Examples


# Requiere que exista el archivo "merluza.csv" en el directorio de trabajo
# Calcula el sexratio por talla y el sexratio por edad
#  sexratio(especie="merluza", edad=T)

# Requiere que exista el archivo "cigala.csv" en el directorio de trabajo
# Calcula el sexratio por talla
#  sexratio(especie="cigala", edad=F)

[Package inbio version 3.0.0 Index]